==> projects/titanic/01_descarga_datos.ipynb <==
# Librerías base

from pathlib import Path
import pandas as pd

url_datos = "https://www.openml.org/data/get_csv/16826755/phpMYEkMl"

datos_titanic = pd.read_csv(url_datos, low_memory=False)

datos_titanic.info()

Las siguientes variables se relacionan con la variable "survived":

* `boat`: El número del bote salvavidas (si sobrevivieron)
* `body`: El número del cuerpo (si no sobrevivieron y el cuerpo fue recuperado)

**Tener presente** un concepto fundamental que se conoce como ***data leakage***, el cual significa que algunas columnas podrían traer consigo la variable que vamos a predecir. Por ejemplo: si conocemos el número del bote salvavidas, esto implica que el pasajero sobrevivió, por lo tanto, la variable "survived" ya se conocería de manera implícita.

columnas_a_usar = [
    "pclass",
    "survived",
    "name",
    "sex",
    "age",
    "sibsp",
    "parch",
    "ticket",
    "fare",
    "cabin",
    "embarked",
    "home.dest",
]

# Para tenerlos en formato csv
datos_titanic_final = pd.read_csv(url_datos, usecols=columnas_a_usar, low_memory=False)

# Verificamos que se hayan cargado las columnas que queríamos
datos_titanic_final.info()

# Obtener la ruta del notebook actual
ruta_datos_crudos = Path.cwd() / "datos"

ruta_archivo = ruta_datos_crudos / "01_datos_crudos_titanic.parquet"

# Crear directorio si no existe
ruta_datos_crudos.mkdir(parents=True, exist_ok=True)

# Guardar el DataFrame en el archivo parquet
datos_titanic_final.to_parquet(ruta_archivo, index=False)
